requirements.txt
astroid==2.12.8
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
flake8==5.0.4
idna==3.3
isort==5.10.1
kaptan==0.5.12
lazy-object-proxy==1.7.1
libtmux==0.14.2
mccabe==0.7.0
platformdirs==2.5.2
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
pylint==2.15.0
PyYAML==5.4.1
requests==2.28.1
tmuxp==1.13.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
